fix: change coins type to float

+++ backend/app/schemas/player_schema.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 class PlayerState(BaseModel):
     username: str
     hp: int = 100
-    coins: int = 10
+    coins: float = 10
     max_units: int = 10
 
     class Config:
